Add basic compressed texture support.
Detects compressed textures by the presence of '.ktx' in the file name.

Not sure if this is the approriate final detection, but it matches the handling for switching the format between gl.RGB | gl.RGBA for PNG alpha channel versus JPG.

An alternative that I considered was having the user always indicate that it was compressed, and what format it was.

Right now, it assumes that all compressed textures will be KTX contained, which will indicate what the internal format is.