Prevent spinner from stalling due to floating point wrap-over.
Some devices, specifically saw on Samsung TVs, seem to map highp floats to be 16 bit. Leading to a frozen spinner after ~65s.

This keeps the uniform as a repeating value through the use of the modulus operator and leaves a little time beyond the period to keep the spinner looking consistent.