fix: Revert original functionality of drawFrame()
Functionality was changed in #402 (Mitigate RAF usage). This is causing grief in our own unit tests as well as unit tests of third party users.

This simple change reverts the original functionality while retaining the new RAF feature.

Here we rename the post-RAF `drawFrame()` method to `renderFrame()`, the name of an existing but verifiably dead method (see below). We then create a new `drawFrame()` which implements its pre-RAF behavior by simply calling `updateFrame()` followed by `renderFrame()`.

The original Stage `renderFrame()` method that is being hijacked here is not used elsewhere in the project (nor the SDK). As for possible third party client uses, this is extremely unlikely given that the `this.ctx.frame` method it calls does not even exist in CoreContext today and would result in an exception if ever attempted. It was removed/renamed by basvanmeurs over 3 years ago (Apr 2019) in 6e0cb2fc0c9335a72bb0b24e578be86ed1acf9eb.

Incidentally, the naming of the functions as `updateFrame()` and `renderFrame()` lines up perfectly with the original function names used in that aforementioned commit.