Fix alpha channel detection when using in-process image handling.
This fix solves an issue when the image worker is not being used, so Lightning processes images on the main thread. There is a simple check for `.png` to toggle whether or not the image has an alpha channel. This however gives an issue when a `.PNG` is used. Introduced a simple lowerCase cast to be able to work with both `.png` and `.PNG`.

Additionally added some (quick and dirty) vitest cases to validate the above changes.
Resolves: https://github.com/rdkcentral/Lightning/issues/493